+++ webapp/src/components/OrderDetail.tsx
@@ -7,9 +7,9 @@ import { dt, t } from "../localization";
 import { dayjs } from "../modules/dayConfig";
 import ScrollTopOnMount from "../shared/ScrollToTopOnMount";
 import Money from "../shared/react/Money";
-import useToggle from "../shared/react/useToggle";
 import useErrorToast from "../state/useErrorToast";
 import useScreenLoader from "../state/useScreenLoader";
+import useToggle from "../state/useToggle";
 import useUser from "../state/useUser";
 import Alert from "../ui/Alert";
 import Button from "../ui/Button";
