+++ webapp/src/components/PrivacyPolicyContent.tsx
@@ -19,8 +19,8 @@ import ScreenLoader from "../components/ScreenLoader";
 import { Lookup, t as loct } from "../localization";
 import { useCurrentLanguage } from "../localization/currentLanguage";
 import useI18n from "../localization/useI18n";
-import useMountEffect from "../shared/react/useMountEffect";
 import useGlobalViewState from "../state/useGlobalViewState";
+import useMountEffect from "../state/useMountEffect";
 import Col from "../ui/Col";
 import Container from "../ui/Container";
 import Nav from "../ui/Nav";
