+++ webapp/src/components/WaitingList.tsx
@@ -4,8 +4,8 @@ import PageLoader from "../components/PageLoader";
 import { t } from "../localization";
 import i18n from "../localization/i18n";
 import useI18n from "../localization/useI18n";
-import useMountEffect from "../shared/react/useMountEffect";
 import useErrorToast from "../state/useErrorToast";
+import useMountEffect from "../state/useMountEffect";
 import useUser from "../state/useUser";
 import Button from "../ui/Button";
 import Form from "../ui/Form";
