+++ webapp/src/localization/I18nProvider.tsx
@@ -4,7 +4,7 @@ import { dayjs } from "../modules/dayConfig";
 import doOnce from "../shared/doOnce";
 import { Logger } from "../shared/logger";
 import { formatMoney } from "../shared/money";
-import useMountEffect from "../shared/react/useMountEffect";
+import useMountEffect from "../state/useMountEffect";
 import useUser from "../state/useUser";
 import { useCurrentLanguage } from "./currentLanguage";
 import i18n from "./i18n";
