+++ webapp/src/pages/Dashboard.tsx
@@ -10,7 +10,7 @@ import { dt, imageAltT, t } from "../localization";
 import { dayjs } from "../modules/dayConfig";
 import externalLinks from "../modules/externalLinks";
 import readOnlyReason from "../modules/readOnlyReason";
-import useAsyncFetch from "../shared/react/useAsyncFetch";
+import useAsyncFetch from "../state/useAsyncFetch";
 import useUser from "../state/useUser";
 import Alert from "../ui/Alert";
 import Button from "../ui/Button";
