+++ webapp/src/pages/FoodCheckoutConfirmation.tsx
@@ -4,7 +4,7 @@ import LayoutContainer from "../components/LayoutContainer";
 import PageLoader from "../components/PageLoader";
 import SumaImage from "../components/SumaImage";
 import { dt, t } from "../localization";
-import useAsyncFetch from "../shared/react/useAsyncFetch";
+import useAsyncFetch from "../state/useAsyncFetch";
 import useUser from "../state/useUser";
 import AlertHeading from "../ui/AlertHeading";
 import Button from "../ui/Button";
