+++ webapp/src/pages/Funding.tsx
@@ -7,10 +7,10 @@ import config from "../config";
 import { t } from "../localization";
 import { dayjs } from "../modules/dayConfig";
 import { scaleMoney } from "../shared/money";
-import useToggle, { Toggle } from "../shared/react/useToggle";
 import useBackendGlobals from "../state/useBackendGlobals";
 import { extractErrorCode, useError } from "../state/useError";
 import useScreenLoader from "../state/useScreenLoader";
+import useToggle, { Toggle } from "../state/useToggle";
 import useUser from "../state/useUser";
 import Button from "../ui/Button";
 import Card from "../ui/Card";
