+++ webapp/src/pages/FundingLinkBankAccount.tsx
@@ -10,8 +10,8 @@ import PageHeading from "../components/PageHeading";
 import config from "../config";
 import { imageAltT, t } from "../localization";
 import keepDigits from "../modules/keepDigits";
-import useHashToggle from "../shared/react/useHashToggle";
 import { extractErrorCode, useError } from "../state/useError";
+import useHashToggle from "../state/useHashToggle";
 import useScreenLoader from "../state/useScreenLoader";
 import useUser from "../state/useUser";
 import Button from "../ui/Button";
