+++ webapp/src/pages/OrderHistoryDetail.tsx
@@ -4,7 +4,7 @@ import ErrorScreen from "../components/ErrorScreen";
 import LayoutContainer from "../components/LayoutContainer";
 import OrderDetail from "../components/OrderDetail";
 import PageLoader from "../components/PageLoader";
-import useAsyncFetch from "../shared/react/useAsyncFetch";
+import useAsyncFetch from "../state/useAsyncFetch";
 import React from "react";
 import { useLocation, useParams } from "react-router-dom";
