+++ webapp/src/pages/PrivateAccountDetail.tsx
@@ -6,10 +6,10 @@ import LayoutContainer from "../components/LayoutContainer";
 import PageLoader from "../components/PageLoader";
 import { dt, t } from "../localization";
 import { scaleMoney } from "../shared/money";
-import useAsyncFetch from "../shared/react/useAsyncFetch";
-import useUnmountEffect from "../shared/react/useUnmountEffect";
+import useAsyncFetch from "../state/useAsyncFetch";
 import { extractErrorCode, useError } from "../state/useError";
 import useScreenLoader from "../state/useScreenLoader";
+import useUnmountEffect from "../state/useUnmountEffect";
 import useUser from "../state/useUser";
 import Alert from "../ui/Alert";
 import Progress from "../ui/Progress";
