+++ webapp/src/pages/PrivateAccountsList.tsx
@@ -7,8 +7,8 @@ import PageLoader from "../components/PageLoader";
 import SumaImage from "../components/SumaImage";
 import { dt, t } from "../localization";
 import ScrollTopOnMount from "../shared/ScrollToTopOnMount";
-import useAsyncFetch from "../shared/react/useAsyncFetch";
-import useMountEffect from "../shared/react/useMountEffect";
+import useAsyncFetch from "../state/useAsyncFetch";
+import useMountEffect from "../state/useMountEffect";
 import Button from "../ui/Button";
 import Card from "../ui/Card";
 import CardBody from "../ui/CardBody";
