+++ webapp/src/pages/RegainAccountAccess.tsx
@@ -4,8 +4,8 @@ import FormControlGroup from "../components/FormControlGroup";
 import FormError from "../components/FormError";
 import PhoneInput from "../components/PhoneInput";
 import { t } from "../localization";
-import useToggle from "../shared/react/useToggle";
 import { extractLocalizedError, useError } from "../state/useError";
+import useToggle from "../state/useToggle";
 import Button from "../ui/Button";
 import Form from "../ui/Form";
 import React from "react";
