+++ webapp/src/pages/Start.tsx
@@ -8,8 +8,8 @@ import { t } from "../localization";
 import useI18n from "../localization/useI18n";
 import { dayjs } from "../modules/dayConfig";
 import { Logger } from "../shared/logger";
-import useToggle from "../shared/react/useToggle";
 import { extractErrorCode, extractLocalizedError, useError } from "../state/useError";
+import useToggle from "../state/useToggle";
 import Form from "../ui/Form";
 import React, { useState } from "react";
 import { useForm } from "react-hook-form";
