+++ webapp/src/pages/Styleguide.tsx
@@ -1,4 +1,4 @@
-import useToggle from "../shared/react/useToggle";
+import useToggle from "../state/useToggle";
 import BrandCard from "../ui/BrandCard";
 import Button from "../ui/Button";
 import ButtonGroup from "../ui/ButtonGroup";
