+++ webapp/src/pages/TripDetail.tsx
@@ -5,7 +5,7 @@ import SumaImage from "../components/SumaImage";
 import { t } from "../localization";
 import { vehicleIconForVendorService } from "../modules/mobilityIconLookup";
 import Money from "../shared/react/Money";
-import useUrlMarshal from "../shared/react/useUrlMarshal";
+import useUrlMarshal from "../state/useUrlMarshal";
 import Stack from "../ui/Stack";
 import dayjs from "dayjs";
 import React from "react";
