+++ webapp/src/pages/Trips.tsx
@@ -5,8 +5,8 @@ import PageLoader from "../components/PageLoader";
 import { t } from "../localization";
 import { vehicleIconForVendorService } from "../modules/mobilityIconLookup";
 import Money from "../shared/react/Money";
-import useAsyncFetch from "../shared/react/useAsyncFetch";
-import useUrlMarshal from "../shared/react/useUrlMarshal";
+import useAsyncFetch from "../state/useAsyncFetch";
+import useUrlMarshal from "../state/useUrlMarshal";
 import Stack from "../ui/Stack";
 import dayjs from "dayjs";
 import isEmpty from "lodash/isEmpty";
